refactor(*.py): Extract utils module. fix #477 @0.5h
