+++ yahooapis.py
@@ -25,30 +25,11 @@ import logging
 import requests
 import requests_toolbelt.adapters.appengine
 
+import utils
+
 # Use the App Engine Requests adapter. This makes sure that Requests uses
 # URLFetch.
 requests_toolbelt.adapters.appengine.monkeypatch()
-
-HIRAGANA_TO_KATAKANA = {
-    u'ぁ': u'ァ', u'あ': u'ア', u'ぃ': u'ィ', u'い': u'イ', u'ぅ': u'ゥ',
-    u'う': u'ウ', u'ぇ': u'ェ', u'え': u'エ', u'ぉ': u'ォ', u'お': u'オ',
-    u'か': u'カ', u'が': u'ガ', u'き': u'キ', u'ぎ': u'ギ', u'く': u'ク',
-    u'ぐ': u'グ', u'け': u'ケ', u'げ': u'ゲ', u'こ': u'コ', u'ご': u'ゴ',
-    u'さ': u'サ', u'ざ': u'ザ', u'し': u'シ', u'じ': u'ジ', u'す': u'ス',
-    u'ず': u'ズ', u'せ': u'セ', u'ぜ': u'ゼ', u'そ': u'ソ', u'ぞ': u'ゾ',
-    u'た': u'タ', u'だ': u'ダ', u'ち': u'チ', u'ぢ': u'ヂ', u'っ': u'ッ',
-    u'つ': u'ツ', u'づ': u'ヅ', u'て': u'テ', u'で': u'デ', u'と': u'ト',
-    u'ど': u'ド', u'な': u'ナ', u'に': u'ニ', u'ぬ': u'ヌ', u'ね': u'ネ',
-    u'の': u'ノ', u'は': u'ハ', u'ば': u'バ', u'ぱ': u'パ', u'ひ': u'ヒ',
-    u'び': u'ビ', u'ぴ': u'ピ', u'ふ': u'フ', u'ぶ': u'ブ', u'ぷ': u'プ',
-    u'へ': u'ヘ', u'べ': u'ベ', u'ぺ': u'ペ', u'ほ': u'ホ', u'ぼ': u'ボ',
-    u'ぽ': u'ポ', u'ま': u'マ', u'み': u'ミ', u'む': u'ム', u'め': u'メ',
-    u'も': u'モ', u'ゃ': u'ャ', u'や': u'ヤ', u'ゅ': u'ュ', u'ゆ': u'ユ',
-    u'ょ': u'ョ', u'よ': u'ヨ', u'ら': u'ラ', u'り': u'リ', u'る': u'ル',
-    u'れ': u'レ', u'ろ': u'ロ', u'ゎ': u'ヮ', u'わ': u'ワ', u'ゐ': u'ヰ',
-    u'ゑ': u'ヱ', u'を': u'ヲ', u'ん': u'ン', u'ゔ': u'ヴ', u'ゕ': u'ヵ',
-    u'ゖ': u'ヶ', u'ゝ': u'ヽ', u'ゞ': u'ヾ',
-}
 
 
 class YahooFuriganaAPI:
@@ -94,22 +75,9 @@ def furigana(yahoo_appid, sentence):
             hiragana += node.firstChild.data
         logging.info(u'furigana:' + hiragana)
 
-        katakana = replace_hiragana_to_katakana(hiragana)
+        katakana = utils.replace_hiragana_to_katakana(hiragana)
         logging.info(u'furigana:' + katakana)
         return katakana
     except Exception as e:
         logging.exception(u'furigana: %s', e)
         return u''
-
-
-def replace_hiragana_to_katakana(hiragana):
-    """ひらがな(平仮名)をカタカナ(片仮名)に変換する
-
-    :param unicode hiragana: ひらがなのテキスト
-    :rtype: unicode
-    :return: カタカナのテキスト
-    """
-    katakana = hiragana
-    for tmp in HIRAGANA_TO_KATAKANA.keys():
-        katakana = katakana.replace(tmp, HIRAGANA_TO_KATAKANA[tmp])
-    return katakana
